test(#664): relocate 4 real-git-spawn suites to integration (C6.1-spawn)
git-source-safety, save-command, walker, and source-providers/git all spawn
a real `git` subprocess (init/commit/push/log) — a genuine process boundary,
not a harness gap. Relocate them out of the unit glob into tests/integration/
(fix relative-import depth) and drop them from ALLOWED_SPAWN, lowering
UNIT_PURITY_BASELINE 32 -> 28.
